pyproject.toml
[project]
name = "gtd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=2.0.0",
    "flet>=0.9.0",
    "inquirerpy>=0.3.4",
    "langchain>=0.0.267",
    "langchain-community>=0.0.1",
    "langchain_openai",
    "mcp[cli]>=1.3.0",
    "openai>=0.27.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "rich>=12.0.0",
    "rumps>=0.4.0 ; sys_platform == 'darwin'",
    "typer>=0.9.0",
    "win10toast>=0.9 ; sys_platform == 'win32'",
]
